Convert and copy image right away if vmt is loaded

+++ src/mainwindow.cpp
@@ -7281,35 +7281,80 @@ void MainWindow::browseVTF( const QString& objectName, QLineEdit* lineEdit ) {
 					texturesToCopy.remove(lineEdit);
 				}
 
-				QString outputFile = fileName.right( fileName.length() - fileName.lastIndexOf('/') - 1 );
+				if(mVMTLoaded) {
+
+					QString newFile = fileName.section("/", -1).section(".", 0, 0);
+					QString dir = QDir::toNativeSeparators(mIniSettings->value("lastSaveAsDir").toString() + "/");
+					QString relativeFilePath = QDir( currentGameMaterialDir() ).relativeFilePath(dir + newFile);
+					Error(dir + newFile + ".vtf");
+
+					if( QFile::exists(dir + newFile + ".vtf") ) {
+						MsgBox msgBox(this);
+						msgBox.setWindowTitle("File already exists!");
+						QPushButton* overwriteButton = msgBox.addButton( "Overwrite", QMessageBox::YesRole );
+						msgBox.addButton( QMessageBox::Cancel );
+						msgBox.setDefaultButton( overwriteButton );
+						msgBox.setIconPixmap(QPixmap(":/icons/info_warning"));;
+
+						msgBox.setText( QDir::toNativeSeparators(dir + newFile + ".vtf") + " already exists. Do you want to overwrite it?"  );
+
+						if(  msgBox.exec() != QMessageBox::Cancel ) {
+
+							if( !QFile::remove( dir + newFile + ".vtf" ) ) {
+								Error( "Error removing \"" + dir + newFile + ".vtf\"" );
+								return;
+							}
+						} else {
+							return;
+						}
+
+						QAction *reconvert = lineEdit->addAction(QIcon(":/icons/reconvert"), QLineEdit::TrailingPosition);
+						lineEdit->setToolTip(fileName);
+						connect(reconvert, SIGNAL(triggered()), SLOT(reconvertTexture()));
+
+						ConversionThread* conversionThread = new ConversionThread(this);
+							conversionThread->fileName = fileName;
+							conversionThread->objectName = objectName;
+							conversionThread->relativeFilePath = relativeFilePath;
+							conversionThread->newFileName = "";
+							conversionThread->outputParameter = "-output \"" + dir + "\"";
+							conversionThread->start();
+
+						lineEdit->setText(relativeFilePath);
+					}
+
+				} else {
+
+					QString outputFile = fileName.right( fileName.length() - fileName.lastIndexOf('/') - 1 );
 					outputFile = outputFile.left( outputFile.indexOf('.') );
 
-				texturesToCopy.insert(lineEdit, outputFile);
+					texturesToCopy.insert(lineEdit, outputFile);
 
-				lineEdit->setText(fileName.right( fileName.length() - fileName.lastIndexOf('/', fileName.lastIndexOf('/') - 1) ));
-				lineEdit->setDisabled(true);
+					lineEdit->setText(fileName.right( fileName.length() - fileName.lastIndexOf('/', fileName.lastIndexOf('/') - 1) ));
+					lineEdit->setDisabled(true);
 
-				QAction *reconvert = lineEdit->addAction(QIcon(":/icons/reconvert"), QLineEdit::TrailingPosition);
-				lineEdit->setToolTip(fileName);
-				connect(reconvert, SIGNAL(triggered()), SLOT(reconvertTexture()));
+					QAction *reconvert = lineEdit->addAction(QIcon(":/icons/reconvert"), QLineEdit::TrailingPosition);
+					lineEdit->setToolTip(fileName);
+					connect(reconvert, SIGNAL(triggered()), SLOT(reconvertTexture()));
 
-				ConversionThread* conversionThread = new ConversionThread(this);
-					conversionThread->fileName = fileName;
-					conversionThread->newFileName = lineEdit->objectName() + "_" + texturesToCopy.value(lineEdit) + ".vtf";
-					conversionThread->outputParameter = "-output \"" + QDir::currentPath().replace("\\", "\\\\") + "\\Cache\\Move\\" + "\"";
-					conversionThread->start();
+					ConversionThread* conversionThread = new ConversionThread(this);
+						conversionThread->fileName = fileName;
+						conversionThread->newFileName = lineEdit->objectName() + "_" + texturesToCopy.value(lineEdit) + ".vtf";
+						conversionThread->outputParameter = "-output \"" + QDir::currentPath().replace("\\", "\\\\") + "\\Cache\\Move\\" + "\"";
+						conversionThread->start();
 
-				fileName.chop(4);
+					fileName.chop(4);
 
-				QString fromFile = fileName.left( fileName.lastIndexOf("/") ) + nameWithExtension;
-				QString toFile = QDir::currentPath() + "/Cache/" + objectName + ".png";
+					QString fromFile = fileName.left( fileName.lastIndexOf("/") ) + nameWithExtension;
+					QString toFile = QDir::currentPath() + "/Cache/" + objectName + ".png";
 
-				if( QFile::exists(toFile) )
-					QFile::remove(toFile);
+					if( QFile::exists(toFile) )
+						QFile::remove(toFile);
 
-				QFile::copy(fromFile, toFile);
+					QFile::copy(fromFile, toFile);
 
-				previewTexture(objectName);
+					previewTexture(objectName);
+				}
 			}
 		}
 	}
@@ -8809,13 +8854,13 @@ void MainWindow::reconvertTexture()
 	else if( objectName == "lineEdit_bump2" )
 		preview = "preview_bumpmap2";
 
-	QString dir = QDir::toNativeSeparators(mIniSettings->value("lastSaveAsDir").toString());
+	QString dir = QDir::toNativeSeparators(mIniSettings->value("lastSaveAsDir").toString() + "/");
 
 	QString fileName = tooltip;
 	QString extension = fileName.section(".", -1);
 	QString newFile = fileName.section("/", -1).section(".", 0, 0);
 
-	QString relativeFilePath = QDir( currentGameMaterialDir() ).relativeFilePath(dir + "/" + newFile);
+	QString relativeFilePath = QDir( currentGameMaterialDir() ).relativeFilePath(dir + newFile);
 
 	if( QFile::exists(dir + newFile + ".vtf") ) {
